# src/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "EDU Science API Document",
    "description": "https://n108.wke.csie.ncnu.edu.tw/",
    "version": "1.0.0"
  },
  "host": "localhost:8888",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/Account/address": {
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "編輯Member地址",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "string": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/Account/phone": {
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "編輯Member電話",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "string": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/Account/birthday": {
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "編輯Member生日",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "string": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/Account/graphql": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "上傳大頭照",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "oid": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/Auth/wkecallback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "WKESSO callback",
        "description": "",
        "parameters": [
          {
            "name": "grantCode",
            "in": "query",
            "type": "string"
          },
          {
            "name": "state",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
    "/Auth/googlecallback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "GoogleSSO callback",
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
    "/Auth/facebookcallback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "FacebookSSO callback",
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
    "/Exam/exam": {
      "get": {
        "tags": [
          "Exam"
        ],
        "summary": "取得所有單一學校所有考卷",
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/Exam/example_exam": {
      "get": {
        "tags": [
          "Exam"
        ],
        "summary": "取得影片相關的題目",
        "description": "",
        "parameters": [
          {
            "name": "oid",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {}
      }
    },
    "/Exam/img_path": {
      "get": {
        "tags": [
          "Exam"
        ],
        "summary": "取得考卷中的題目",
        "description": "",
        "parameters": [
          {
            "name": "cid",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {}
      }
    },
    "/Exam/issue": {
      "put": {
        "tags": [
          "Exam"
        ],
        "summary": "提題目交issue",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "pid": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
  },
...
}